agent join and stream api

+++ src/ConsulSharp/V1/Agent/CoordinateModel.cs
@@ -5,10 +5,10 @@ namespace ConsulSharp.V1.ACL.Agent
 {
     public class CoordinateModel
     {
-        public int Adjustment { get; set; }
+        public double Adjustment { get; set; }
         public double Error { get; set; }
         
         [JsonProperty("Vec")]
-        public List<int> Vectors { get; set; }
+        public List<double> Vectors { get; set; }
     }
 }
