+++ src/ConsulSharp/V1/Agent/Counter.cs
@@ -5,12 +5,12 @@ namespace ConsulSharp.V1.ACL.Agent
     public class Counter
     {
         public string Name { get; set; }
-        public int Count { get; set; }
-        public int Sum { get; set; }
-        public int Min { get; set; }
-        public int Max { get; set; }
-        public int Mean { get; set; }
-        public int Stddev { get; set; }
+        public double Count { get; set; }
+        public double Sum { get; set; }
+        public double Min { get; set; }
+        public double Max { get; set; }
+        public double Mean { get; set; }
+        public double Stddev { get; set; }
 
         public Dictionary<string, object> Labels { get; set; }
 }
